requirements.txt
streamlit==1.31.0
pandas==2.2.0
numpy==1.26.3
google-generativeai==0.3.2
plotly==5.18.0
seaborn==0.13.1
matplotlib==3.8.2
scipy==1.12.0
python-dotenv==1.0.0